# Performance tests.