--- a/examples/Benchmark/HTTP/Makefile
+++ b/examples/Benchmark/HTTP/Makefile
@@ -1,0 +1,1 @@
+../../../scripts/Makefile