--- a/examples/Benchmark/HTTP/README.md
+++ b/examples/Benchmark/HTTP/README.md
@@ -1,0 +1,3 @@
+## `Benchmark/HTTP`
+
+A simple "A+B over HTTP" benchmark. 20+QPS on our "golden" Hetzner instance. -- D.K.