--- a/examples/Benchmark/HTTP/server.h
+++ b/examples/Benchmark/HTTP/server.h
@@ -40,7 +40,9 @@ class BenchmarkTestServer {
                                    [](Request r) {
                                      r(AddResult(current::FromString<int64_t>(r.url.query["a"]) +
                                                  current::FromString<int64_t>(r.url.query["b"])));
-                                   })) {}
+                                   }) +
+              HTTP(port).Register("/perftest",
+                                   [](Request r) { r("perftest ok\n");})) {}
 
   void Join() { HTTP(port_).Join(); }