--- a/examples/Benchmark/generic/Makefile
+++ b/examples/Benchmark/generic/Makefile
@@ -1,0 +1,1 @@
+../../../scripts/Makefile