--- a/examples/Benchmark/generic/README.md
+++ b/examples/Benchmark/generic/README.md
@@ -1,0 +1,10 @@
+## `Benchmark/Primes`
+
+The "is a random number between one and one million prime" benchmark, comparing:
+
+* local, no-HTTP, run
+* simple HTTP server,
+* a `Storage`-based solution, and
+* a `Storage`-based solution with "authentication".
+
+TODO(dkorolev): Run instructions.