The actual commit: https://github.com/C5T/Current/commit/1de35cf3cc0a75a8228214c3f1eaa258db74faa1

(As this PR incorporates https://github.com/C5T/Current/pull/506 as of now).

Max, I'm seeing 33K QPS on Current's HTTP server on Hetz, and 16K..18K on my T420s. Fucking impressive.

At the same time, an attempt to test the QPS of a plain nginx-based `return 200 "OK"` fails for the reason I don't understand. Looks like nginx config update does go through, but sometimes fails to stick or something.

On the other hand, the issue may be with Current's HTTP client, somehow bringing the port into some unresponsive state. Here's what I can reproduce even for no-nginx test:

```
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
16633.6 QPS.
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
18117.6 QPS.
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
16392.8 QPS.
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
16229.2 QPS.
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
terminate called after throwing an instance of 'current::net::SocketConnectException'
  what():  ./../../../Sherlock/../Blocks/HTTP/../../Bricks/net/http/impl/../../tcp/impl/posix.h:560	SocketConnectException()
Aborted (core dumped)
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ 
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
terminate called after throwing an instance of 'current::net::SocketConnectException'
  what():  ./../../../Sherlock/../Blocks/HTTP/../../Bricks/net/http/impl/../../tcp/impl/posix.h:560	SocketConnectException()
Aborted (core dumped)
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
terminate called after throwing an instance of 'current::net::SocketConnectException'
  what():  ./../../../Sherlock/../Blocks/HTTP/../../Bricks/net/http/impl/../../tcp/impl/posix.h:560	SocketConnectException()
Aborted (core dumped)
```

Could you try on your Mac please?

Thanks,
Dima