@dkorolev 2016-05-23
Just in case: when carefully run with nginx configured manually, Current's HTTP server is on par with nginx.

```
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ git diff --no-ext-diff 
diff --git a/examples/Benchmark/generic/scenario_nginx_client.h b/examples/Benchmark/generic/scenario_nginx_client.h
index 2ceab7c..f3068cf 100644
--- a/examples/Benchmark/generic/scenario_nginx_client.h
+++ b/examples/Benchmark/generic/scenario_nginx_client.h
@@ -63,7 +63,7 @@ SCENARIO(simple_nginx_client, "Use Current's HTTP client to access an nginx-cont
     server_directive.CreateLocation(FLAGS_nginx_client_local_route)
         .Add(SimpleDirective("return", "200 \"" + FLAGS_nginx_client_test_body + "\""));
 
-    nginx_manager_.UpdateConfig(std::move(server_directive));
+    //nginx_manager_.UpdateConfig(std::move(server_directive));
   }
 
   void RunOneQuery() override { assert(HTTP(GET(url_)).body == FLAGS_nginx_client_test_body); }
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ cat /etc/nginx/sites-enabled/current_perftest
server {
  listen 11111;
  location /perftest {
    return 200 "200 OK generated by nginx!";
  }
  location /ae00acdc38ead2003665d84d2dd22a7e9de58b955ac9509bea2f45643f83809f {
    return 200 "ae2fb999cd61669524c909fc12d9000646d348484ddeffc84550f451f099e446";
  }
}
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ curl localhost:11111/perftest && echo
200 OK generated by nginx!
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=simple_nginx_client
17043.6 QPS.
dima@dima-i9: ~/github/dkorolev/Current/examples/Benchmark/generic(perftest) $ ./.current/run --scenario=current_http_server
17904 QPS.
```

Tested with less threads too. `nginx`'s QPS is higher on `--threads=1` and `--threads=4`, but it's already even with `threads` >= 20.