@dkorolev 2016-05-23
And confirmed the length of the message from a `Current`-powered HTTP server does not affect performance on my local machine.