@dkorolev 2016-05-23
@mzhurovich , https://github.com/C5T/Current/pull/507/commits/a244ded50f47df139d95a8e2af7305ac1c97c95c introduces port spanning.

Current's HTTP server is officially kicking nginx's ass ... on some machines, but still.

Thanks,
Dima